Separated commands into intake and arm folders
