+++ src/main/java/frc/robot/commands/arm/ArmManual.java
@@ -2,24 +2,31 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.intake;
+package frc.robot.commands.arm;
 
+import java.util.function.DoubleSupplier;
+
+import edu.wpi.first.math.filter.SlewRateLimiter;
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.subsystems.arm.ArmSubsystem;
 
 /* You should consider using the more terse Command factories API instead https://docs.wpilib.org/en/stable/docs/software/commandbased/organizing-command-based.html#defining-commands */
-public class ArmHold extends Command {
+public class ArmManual extends Command {
 
   private ArmSubsystem armSub;
-  private double angle;
+  private DoubleSupplier speed;
 
-  /** Creates a new ArmHold. */
-  public ArmHold(ArmSubsystem armSub) {
+  SlewRateLimiter filter = new SlewRateLimiter(0.9);
+
+  /** Creates a new PivotManual. */
+  public ArmManual(ArmSubsystem armSub, DoubleSupplier d) {
     // Use addRequirements() here to declare subsystem dependencies.
 
     this.armSub = armSub;
-    angle = armSub.getArmPosition();
+    this.speed = d;
+
     addRequirements(armSub);
+
   }
 
   // Called when the command is initially scheduled.
@@ -30,12 +37,17 @@ public class ArmHold extends Command {
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    armSub.armHold(angle);
+    if (Math.abs(speed.getAsDouble()) < 0.3) {
+      armSub.setArmSpeed(speed.getAsDouble());
+    } else {
+      armSub.setArmSpeed(filter.calculate(speed.getAsDouble()) * 0.7);
+    }
   }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
+    armSub.setArmSpeed(0);
   }
 
   // Returns true when the command should end.
